+++ client/index.html
@@ -11,7 +11,7 @@
     <script type="module" src="/src/main.jsx"></script>
   </body>
 
-  <script src="/socket.io/socket.io.js"></script>
+  <script type="module" src="/socket.io/socket.io.js"></script>
   <script>
     var socket = io();
   </script>
